// src/common/interfaces/DTO/BarriersEvaluationDTO.ts
export interface BarriersEvaluationDTO {
    ref_id?: string;
    barrier_id?: string;
    status?: string;
    barrier_element_id?: string;
    barrier_envelope_id?: string;
    barrier_diagram_id?: string;
    scenario_id?: string;
    create_user?: string;
    component_ovality?: number;
    component_wearing?: number;
    details:string;
  }
  